+++ plugins/verilog_parser/include/verilog_parser/verilog_module.h
@@ -113,6 +113,7 @@ namespace hal
         std::vector<std::unique_ptr<VerilogPort>> m_ports;
         std::map<std::string, VerilogPort*> m_ports_by_identifier;
         std::map<std::string, VerilogPort*> m_ports_by_expression;
+        std::set<std::string> m_expanded_port_expressions;
 
         // signals
         std::vector<std::unique_ptr<VerilogSignal>> m_signals;